+++ apps/doc/src/app/about-prizm/changelog/CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Features
 
+- feat(components/breadcrumbs): added ability to work with projections #464
 - feat(components/overlay): added new logic for detecting events from dynamic elements #411
 - feat(helpers/rxjs): 'raceEmit' operator was added to detect first emit in the transmitted time interval between source streams